Move the Display models into the new package
