+++ catalogue_api/api/src/main/scala/uk/ac/wellcome/platform/api/models/DisplayResultList.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.platform.api.models
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.swagger.annotations.{ApiModel, ApiModelProperty}
 import uk.ac.wellcome.models.WorksIncludes
-import uk.ac.wellcome.models.display.DisplayWork
+import uk.ac.wellcome.display.models.DisplayWork
 
 @ApiModel(
   value = "ResultList",
