+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/services/WorksServiceTest.scala
@@ -5,7 +5,7 @@ import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.{IdentifierSchemes, SourceIdentifier}
 import uk.ac.wellcome.platform.api.WorksUtil
 import uk.ac.wellcome.models.WorksIncludes
-import uk.ac.wellcome.models.display.{DisplayIdentifier, DisplayWork}
+import uk.ac.wellcome.display.models.{DisplayIdentifier, DisplayWork}
 import uk.ac.wellcome.platform.api.models.DisplayResultList
 import uk.ac.wellcome.test.utils.IndexedElasticSearchLocal
