+++ sbt_common/display/src/test/scala/uk/ac/wellcome/models/display/DisplayLicenseTest.scala
